requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# Multi-Agent Orchestration
langgraph==0.2.38
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-anthropic==0.2.4
langchain-openai==0.2.5
anthropic==0.36.2

# Deep Learning - Script Detection
tensorflow==2.17.0
keras==3.5.0
torch==2.4.1
torchvision==0.19.1

# Image Processing & OCR
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
scikit-image==0.24.0

# OCR Models
transformers==4.45.2
datasets==3.0.1
timm==1.0.11
sentencepiece==0.2.0

# Text Detection
ultralytics==8.3.0

# Image Restoration
basicsr==1.4.2
realesrgan==0.3.0

# RAG / Vector DB
chromadb==0.5.15
sentence-transformers==3.2.1
faiss-cpu==1.9.0

# Task Queue
celery==5.4.0
redis==5.1.1
flower==2.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
alembic==1.13.3

# API Utilities
httpx==0.27.2
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
prometheus-client==0.21.0
structlog==24.4.0
sentry-sdk[fastapi]==2.16.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-mock==3.14.0

# Utilities
python-dotenv==1.0.1
boto3==1.35.32
tqdm==4.66.5
PyYAML==6.0.2
